@@ -4,22 +4,22 @@ import (
 	"crypto/tls"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"net/url"
 	"strconv"
 	"strings"
 	"time"
-	"unicode/utf8"
 
 	"github.com/corpix/uarand"
 	"github.com/microcosm-cc/bluemonday"
 	"github.com/projectdiscovery/cdncheck"
 	"github.com/projectdiscovery/fastdialer/fastdialer"
-	pdhttputil "github.com/projectdiscovery/httputil"
+	"github.com/projectdiscovery/gologger"
 	"github.com/projectdiscovery/rawhttp"
 	retryablehttp "github.com/projectdiscovery/retryablehttp-go"
-	"github.com/projectdiscovery/stringsutil"
+	pdhttputil "github.com/projectdiscovery/utils/http"
+	stringsutil "github.com/projectdiscovery/utils/strings"
+	"golang.org/x/net/context"
 	"golang.org/x/net/http2"
 )
 
@@ -42,6 +42,12 @@ func New(options *Options) (*HTTPX, error) {
 	fastdialerOpts.EnableFallback = true
 	fastdialerOpts.Deny = options.Deny
 	fastdialerOpts.Allow = options.Allow
+	fastdialerOpts.WithDialerHistory = true
+	fastdialerOpts.WithZTLS = options.ZTLS
+	if len(options.Resolvers) > 0 {
+		fastdialerOpts.BaseResolvers = options.Resolvers
+	}
+	fastdialerOpts.SNIName = options.SniName
 	dialer, err := fastdialer.NewDialer(fastdialerOpts)
 	if err != nil {
 		return nil, fmt.Errorf("could not create resolver cache: %s", err)
@@ -49,6 +55,8 @@ func New(options *Options) (*HTTPX, error) {
 	httpx.Dialer = dialer
 
 	httpx.Options = options
+
+	httpx.Options.parseCustomCookies()
 
 	var retryablehttpOptions = retryablehttp.DefaultOptionsSpraying
 	retryablehttpOptions.Timeout = httpx.Options.Timeout
@@ -62,6 +70,8 @@ func New(options *Options) (*HTTPX, error) {
 	if httpx.Options.FollowRedirects {
 		// Follow redirects up to a maximum number
 		redirectFunc = func(redirectedRequest *http.Request, previousRequests []*http.Request) error {
+			// add custom cookies if necessary
+			httpx.setCustomCookies(redirectedRequest)
 			if len(previousRequests) >= options.MaxRedirects {
 				// https://github.com/golang/go/issues/10069
 				return http.ErrUseLastResponse
@@ -73,9 +83,15 @@ func New(options *Options) (*HTTPX, error) {
 	if httpx.Options.FollowHostRedirects {
 		// Only follow redirects on the same host up to a maximum number
 		redirectFunc = func(redirectedRequest *http.Request, previousRequests []*http.Request) error {
-			// Check if we get a redirect to a differen host
+			// add custom cookies if necessary
+			httpx.setCustomCookies(redirectedRequest)
+
+			// Check if we get a redirect to a different host
 			var newHost = redirectedRequest.URL.Host
-			var oldHost = previousRequests[0].URL.Host
+			var oldHost = previousRequests[0].Host
+			if oldHost == "" {
+				oldHost = previousRequests[0].URL.Host
+			}
 			if newHost != oldHost {
 				// Tell the http client to not follow redirect
 				return http.ErrUseLastResponse
@@ -87,14 +103,19 @@ func New(options *Options) (*HTTPX, error) {
 			return nil
 		}
 	}
-
 	transport := &http.Transport{
 		DialContext:         httpx.Dialer.Dial,
+		DialTLSContext:      httpx.Dialer.DialTLS,
 		MaxIdleConnsPerHost: -1,
 		TLSClientConfig: &tls.Config{
 			InsecureSkipVerify: true,
+			MinVersion:         tls.VersionTLS10,
 		},
 		DisableKeepAlives: true,
+	}
+
+	if httpx.Options.SniName != "" {
+		transport.TLSClientConfig.ServerName = httpx.Options.SniName
 	}
 
 	if httpx.Options.HTTPProxy != "" {
@@ -111,14 +132,19 @@ func New(options *Options) (*HTTPX, error) {
 		CheckRedirect: redirectFunc,
 	}, retryablehttpOptions)
 
-	httpx.client2 = &http.Client{
-		Transport: &http2.Transport{
-			TLSClientConfig: &tls.Config{
-				InsecureSkipVerify: true,
-			},
-			AllowHTTP: true,
+	transport2 := &http2.Transport{
+		TLSClientConfig: &tls.Config{
+			InsecureSkipVerify: true,
+			MinVersion:         tls.VersionTLS10,
 		},
-		Timeout: httpx.Options.Timeout,
+		AllowHTTP: true,
+	}
+	if httpx.Options.SniName != "" {
+		transport2.TLSClientConfig.ServerName = httpx.Options.SniName
+	}
+	httpx.client2 = &http.Client{
+		Transport: transport2,
+		Timeout:   httpx.Options.Timeout,
 	}
 
 	httpx.htmlPolicy = bluemonday.NewPolicy()
@@ -126,7 +152,7 @@ func New(options *Options) (*HTTPX, error) {
 	if options.CdnCheck || options.ExcludeCdn {
 		httpx.cdn, err = cdncheck.NewWithCache()
 		if err != nil {
-			return nil, fmt.Errorf("could not create cdn check: %s", err)
+			gologger.Error().Msgf("could not create cdn check: %v", err)
 		}
 	}
 
@@ -140,13 +166,13 @@ func (h *HTTPX) Do(req *retryablehttp.Request, unsafeOptions UnsafeOptions) (*Re
 	var gzipRetry bool
 get_response:
 	httpresp, err := h.getResponse(req, unsafeOptions)
-	if err != nil {
+	if httpresp == nil && err != nil {
 		return nil, err
 	}
 
 	var shouldIgnoreErrors, shouldIgnoreBodyErrors bool
 	switch {
-	case h.Options.Unsafe && req.Method == http.MethodHead && !stringsutil.ContainsAny("i/o timeout"):
+	case h.Options.Unsafe && req.Method == http.MethodHead && !stringsutil.ContainsAny(err.Error(), "i/o timeout"):
 		shouldIgnoreErrors = true
 		shouldIgnoreBodyErrors = true
 	}
@@ -158,6 +184,11 @@ get_response:
 	// httputil.DumpResponse does not handle websockets
 	headers, rawResp, err := pdhttputil.DumpResponseHeadersAndRaw(httpresp)
 	if err != nil {
+		if stringsutil.ContainsAny(err.Error(), "tls: user canceled") {
+			shouldIgnoreErrors = true
+			shouldIgnoreBodyErrors = true
+		}
+
 		// Edge case - some servers respond with gzip encoding header but uncompressed body, in this case the standard library configures the reader as gzip, triggering an error when read.
 		// The bytes slice is not accessible because of abstraction, therefore we need to perform the request again tampering the Accept-Encoding header
 		if !gzipRetry && strings.Contains(err.Error(), "gzip: invalid header") {
@@ -171,12 +202,11 @@ get_response:
 	}
 	resp.Raw = string(rawResp)
 	resp.RawHeaders = string(headers)
-
 	var respbody []byte
 	// websockets don't have a readable body
 	if httpresp.StatusCode != http.StatusSwitchingProtocols {
 		var err error
-		respbody, err = ioutil.ReadAll(io.LimitReader(httpresp.Body, h.Options.MaxResponseBodySizeToRead))
+		respbody, err = io.ReadAll(io.LimitReader(httpresp.Body, h.Options.MaxResponseBodySizeToRead))
 		if err != nil && !shouldIgnoreBodyErrors {
 			return nil, err
 		}
@@ -187,6 +217,11 @@ get_response:
 		return nil, closeErr
 	}
 
+	respbody, err = DecodeData(respbody, httpresp.Header)
+	if err != nil && !shouldIgnoreBodyErrors {
+		return nil, closeErr
+	}
+
 	respbodystr := string(respbody)
 
 	// check if we need to strip html
@@ -194,12 +229,17 @@ get_response:
 		respbodystr = h.htmlPolicy.Sanitize(respbodystr)
 	}
 
-	if contentLength, ok := resp.Headers["Content-Length"]; ok {
-		contentLengthInt, err := strconv.Atoi(strings.Join(contentLength, ""))
-		if err != nil {
-			resp.ContentLength = utf8.RuneCountInString(respbodystr)
-		} else {
+	// if content length is not defined
+	if resp.ContentLength <= 0 {
+		// check if it's in the header and convert to int
+		if contentLength, ok := resp.Headers["Content-Length"]; ok && len(contentLength) > 0 {
+			contentLengthInt, _ := strconv.Atoi(contentLength[0])
 			resp.ContentLength = contentLengthInt
+		}
+
+		// if we have a body, then use the number of bytes in the body if the length is still zero
+		if resp.ContentLength <= 0 && len(respbody) > 0 {
+			resp.ContentLength = len(respbody)
 		}
 	}
 
@@ -217,7 +257,7 @@ get_response:
 		resp.TLSData = h.TLSGrab(httpresp)
 	}
 
-	resp.CSPData = h.CSPGrab(httpresp)
+	resp.CSPData = h.CSPGrab(&resp)
 
 	// build the redirect flow by reverse cycling the response<-request chain
 	if !h.Options.Unsafe {
@@ -241,11 +281,10 @@ type UnsafeOptions struct {
 }
 
 // getResponse returns response from safe / unsafe request
-func (h *HTTPX) getResponse(req *retryablehttp.Request, unsafeOptions UnsafeOptions) (*http.Response, error) {
+func (h *HTTPX) getResponse(req *retryablehttp.Request, unsafeOptions UnsafeOptions) (resp *http.Response, err error) {
 	if h.Options.Unsafe {
 		return h.doUnsafeWithOptions(req, unsafeOptions)
 	}
-
 	return h.client.Do(req)
 }
 
@@ -288,7 +327,12 @@ func (h *HTTPX) AddFilter(f Filter) {
 
 // NewRequest from url
 func (h *HTTPX) NewRequest(method, targetURL string) (req *retryablehttp.Request, err error) {
-	req, err = retryablehttp.NewRequest(method, targetURL, nil)
+	return h.NewRequestWithContext(context.Background(), method, targetURL)
+}
+
+// NewRequest from url
+func (h *HTTPX) NewRequestWithContext(ctx context.Context, method, targetURL string) (req *retryablehttp.Request, err error) {
+	req, err = retryablehttp.NewRequestWithContext(ctx, method, targetURL, nil)
 	if err != nil {
 		return
 	}
@@ -306,13 +350,25 @@ func (h *HTTPX) NewRequest(method, targetURL string) (req *retryablehttp.Request
 // SetCustomHeaders on the provided request
 func (h *HTTPX) SetCustomHeaders(r *retryablehttp.Request, headers map[string]string) {
 	for name, value := range headers {
-		r.Header.Set(name, value)
-		// host header is particular
-		if strings.EqualFold(name, "host") {
+		switch strings.ToLower(name) {
+		case "host":
 			r.Host = value
+		case "cookie":
+			// cookies are set in the default branch, and reset during the follow redirect flow
+			fallthrough
+		default:
+			r.Header.Set(name, value)
 		}
 	}
 	if h.Options.RandomAgent {
 		r.Header.Set("User-Agent", uarand.GetRandom()) //nolint
+	}
+}
+
+func (httpx *HTTPX) setCustomCookies(req *http.Request) {
+	if httpx.Options.hasCustomCookies() {
+		for _, cookie := range httpx.Options.customCookies {
+			req.AddCookie(cookie)
+		}
 	}
 }
