@@ -2,6 +2,7 @@ package main
 
 import (
 	"fmt"
+	"io"
 	"net/http"
 	"net/http/httptest"
 	"strings"
@@ -11,20 +12,32 @@ import (
 )
 
 var httpTestcases = map[string]testutils.TestCase{
-	"Standard HTTP GET Request":                                                 &standardHttpGet{},
-	"Standard HTTPS GET Request":                                                &standardHttpGet{tls: true},
-	"Raw HTTP GET Request":                                                      &standardHttpGet{unsafe: true},
-	"Raw request with non standard rfc path via stdin":                          &standardHttpGet{unsafe: true, stdinPath: "/%invalid"},
-	"Raw request with non standard rfc path via cli flag":                       &standardHttpGet{unsafe: true, path: "/%invalid"},
-	"Regression test for: https://github.com/projectdiscovery/httpx/issues/363": &issue363{}, // infinite redirect
-	"Regression test for: https://github.com/projectdiscovery/httpx/issues/276": &issue276{}, // full path with port in output
-	"Regression test for: https://github.com/projectdiscovery/httpx/issues/277": &issue277{}, // scheme://host:port via stdin
-	"Regression test for: https://github.com/projectdiscovery/httpx/issues/303": &issue303{}, // misconfigured gzip header with uncompressed body
+	"Standard HTTP GET Request":                                                           &standardHttpGet{},
+	"Standard HTTPS GET Request":                                                          &standardHttpGet{tls: true},
+	"Raw HTTP GET Request":                                                                &standardHttpGet{unsafe: true},
+	"Raw request with non standard rfc path via stdin":                                    &standardHttpGet{unsafe: true, stdinPath: "/%invalid"},
+	"Raw request with non standard rfc path via cli flag":                                 &standardHttpGet{unsafe: true, path: "/%invalid"},
+	"Regression test for: https://github.com/projectdiscovery/httpx/issues/363":           &issue363{}, // infinite redirect
+	"Regression test for: https://github.com/projectdiscovery/httpx/issues/276":           &issue276{}, // full path with port in output
+	"Regression test for: https://github.com/projectdiscovery/httpx/issues/277":           &issue277{}, // scheme://host:port via stdin
+	"Regression test for: https://github.com/projectdiscovery/httpx/issues/303":           &issue303{}, // misconfigured gzip header with uncompressed body
+	"Regression test for: https://github.com/projectdiscovery/httpx/issues/400":           &issue400{}, // post operation with body
+	"Regression test for: https://github.com/projectdiscovery/httpx/issues/414":           &issue414{}, // stream mode with path
+	"Regression test for: https://github.com/projectdiscovery/httpx/issues/433":           &issue433{}, // new line scanning with title flag
+	"Request URI to existing file - https://github.com/projectdiscovery/httpx/issues/480": &issue480{}, // request uri pointing to existing file
+	"Standard HTTP GET Request with match response time":                                  &standardHttpGet{mrt: true, inputValue: "\"<10s\""},
+	"Standard HTTP GET Request with filter response time":                                 &standardHttpGet{frt: true, inputValue: "\">3ms\""},
+	"Multiple Custom Header":                                                              &customHeader{inputData: []string{"-debug-req", "-H", "'user-agent: test'", "-H", "'foo: bar'"}, expectedOutput: []string{"User-Agent: test", "Foo: bar"}},
+	"Output Match Condition":                                                              &outputMatchCondition{inputData: []string{"-silent", "-mdc", "\"status_code == 200\""}},
+	"Output Filter Condition":                                                             &outputFilterCondition{inputData: []string{"-silent", "-fdc", "\"status_code == 400\""}},
 }
 
 type standardHttpGet struct {
 	tls            bool
 	unsafe         bool
+	mrt            bool
+	frt            bool
+	inputValue     string
 	stdinPath      string
 	path           string
 	expectedOutput string
@@ -50,7 +63,12 @@ func (h *standardHttpGet) Execute() error {
 	if h.path != "" {
 		extra = append(extra, "-path", "\""+h.path+"\"")
 	}
-
+	if h.mrt {
+		extra = append(extra, "-mrt", h.inputValue)
+	}
+	if h.frt {
+		extra = append(extra, "-frt", h.inputValue)
+	}
 	URL := ts.URL
 	if h.stdinPath != "" {
 		URL += h.stdinPath
@@ -65,7 +83,7 @@ func (h *standardHttpGet) Execute() error {
 	}
 
 	if h.expectedOutput != "" && !strings.EqualFold(results[0], h.expectedOutput) {
-		return errIncorrectResult(results[0], h.expectedOutput)
+		return errIncorrectResult(h.expectedOutput, results[0])
 	}
 
 	return nil
@@ -97,7 +115,7 @@ func (h *issue276) Execute() error {
 	// title
 	expected := ts.URL + "/redirect" + " [302] [Object moved]"
 	if !strings.EqualFold(results[0], expected) {
-		return errIncorrectResult(results[0], expected)
+		return errIncorrectResult(expected, results[0])
 	}
 	return nil
 }
@@ -128,7 +146,7 @@ func (h *issue277) Execute() error {
 	// title
 	expected := ts.URL + uripath
 	if !strings.EqualFold(results[0], expected) {
-		return errIncorrectResult(results[0], expected)
+		return errIncorrectResult(expected, results[0])
 	}
 	return nil
 }
@@ -157,7 +175,7 @@ func (h *issue303) Execute() error {
 	// full url with port
 	expected := ts.URL
 	if !strings.EqualFold(results[0], expected) {
-		return errIncorrectResult(results[0], expected)
+		return errIncorrectResult(expected, results[0])
 	}
 	return nil
 }
@@ -179,6 +197,181 @@ func (h *issue363) Execute() error {
 	if err != nil {
 		return err
 	}
+	if len(results) != 1 {
+		return errIncorrectResultsCount(results)
+	}
+	return nil
+}
+
+type issue400 struct{}
+
+func (h *issue400) Execute() error {
+	var ts *httptest.Server
+	router := httprouter.New()
+	router.POST("/receive", httprouter.Handle(func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+		w.Header().Add("Content-Type", "application/json")
+		data, _ := io.ReadAll(r.Body)
+		fmt.Fprintf(w, "data received %s", data)
+	}))
+	ts = httptest.NewServer(router)
+	defer ts.Close()
+
+	results, err := testutils.RunHttpxAndGetResults(ts.URL+"/receive", debug, "-body 'a=b'", "-x POST", "-status-code")
+	if err != nil {
+		return err
+	}
+	if len(results) != 1 {
+		return errIncorrectResultsCount(results)
+	}
+	return nil
+}
+
+type issue414 struct{}
+
+func (h *issue414) Execute() error {
+	var ts *httptest.Server
+	uripath := "/path"
+	router := httprouter.New()
+	router.POST(uripath, httprouter.Handle(func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+		w.Header().Add("Content-Type", "application/json")
+		data, _ := io.ReadAll(r.Body)
+		fmt.Fprintf(w, "data received %s", data)
+	}))
+	ts = httptest.NewServer(router)
+	defer ts.Close()
+
+	results, err := testutils.RunHttpxAndGetResults(ts.URL, debug, fmt.Sprintf("-path '%s'", uripath))
+	if err != nil {
+		return err
+	}
+	if len(results) != 1 {
+		return errIncorrectResultsCount(results)
+	}
+	expected := ts.URL + uripath
+	if !strings.EqualFold(results[0], expected) {
+		return errIncorrectResult(expected, results[0])
+	}
+	return nil
+}
+
+type issue433 struct{}
+
+func (h *issue433) Execute() error {
+	var ts *httptest.Server
+	router := httprouter.New()
+	uriPath := "/index"
+	router.GET(uriPath, httprouter.Handle(func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+		htmlResponse := "<html><head><title>Project\n\r Discovery\n - Httpx></title></head><body>test data</body></html>"
+		fmt.Fprint(w, htmlResponse)
+	}))
+	ts = httptest.NewServer(router)
+	defer ts.Close()
+	results, err := testutils.RunHttpxAndGetResults(fmt.Sprint(ts.URL, uriPath), debug, "-title", "-no-color")
+	if err != nil {
+		return err
+	}
+	if strings.Contains(results[0], "\n") {
+		return errIncorrectResultsCount(results)
+	}
+	if strings.Contains(results[0], "\r") {
+		return errIncorrectResultsCount(results)
+	}
+	return nil
+}
+
+type issue480 struct{}
+
+func (h *issue480) Execute() error {
+	var ts *httptest.Server
+	router := httprouter.New()
+	uriPath := "////////////////../../../../../../../../etc/passwd"
+	router.GET(uriPath, httprouter.Handle(func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+		htmlResponse := "<html><body>ok from uri</body></html>"
+		fmt.Fprint(w, htmlResponse)
+	}))
+	ts = httptest.NewServer(router)
+	defer ts.Close()
+	results, err := testutils.RunHttpxAndGetResults(ts.URL, debug, "-path", "////////////////../../../../../../../../etc/passwd")
+	if err != nil {
+		return err
+	}
+	if !strings.Contains(results[0], uriPath) {
+		return errIncorrectResultsCount(results)
+	}
+	return nil
+}
+
+type customHeader struct {
+	inputData      []string
+	expectedOutput []string
+}
+
+func (h *customHeader) Execute() error {
+	var ts *httptest.Server
+	router := httprouter.New()
+	router.GET("/", httprouter.Handle(func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+		w.Header().Add("Content-Type", "application/json")
+		fmt.Fprint(w, `{"status": "ok"}`)
+	}))
+	ts = httptest.NewServer(router)
+	defer ts.Close()
+
+	results, err := testutils.RunHttpxAndGetCombinedResults(ts.URL, true, h.inputData...)
+	if err != nil {
+		return err
+	}
+	for _, expected := range h.expectedOutput {
+		if !strings.Contains(results, expected) {
+			return errIncorrectResult(expected, results)
+		}
+	}
+	return nil
+}
+
+type outputMatchCondition struct {
+	inputData []string
+}
+
+func (h *outputMatchCondition) Execute() error {
+	var ts *httptest.Server
+	router := httprouter.New()
+	router.GET("/", httprouter.Handle(func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+		w.Header().Add("Content-Type", "application/json")
+		w.WriteHeader(200)
+		fmt.Fprint(w, `{"status": "ok"}`)
+	}))
+	ts = httptest.NewServer(router)
+	defer ts.Close()
+	results, err := testutils.RunHttpxAndGetResults(ts.URL, false, h.inputData...)
+	if err != nil {
+		return err
+	}
+	if len(results) != 1 {
+		return errIncorrectResultsCount(results)
+	}
+	return nil
+}
+
+type outputFilterCondition struct {
+	inputData []string
+}
+
+func (h *outputFilterCondition) Execute() error {
+	var ts *httptest.Server
+	router := httprouter.New()
+	router.GET("/", httprouter.Handle(func(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+		w.Header().Add("Content-Type", "application/json")
+		w.WriteHeader(200)
+		fmt.Fprint(w, `{"status": "ok"}`)
+	}))
+	ts = httptest.NewServer(router)
+	defer ts.Close()
+
+	results, err := testutils.RunHttpxAndGetResults(ts.URL, false, h.inputData...)
+	if err != nil {
+		return err
+	}
+
 	if len(results) != 1 {
 		return errIncorrectResultsCount(results)
 	}
