@@ -2,6 +2,8 @@ package customlist
 
 import "github.com/projectdiscovery/httpx/common/fileutil"
 
+const maxRecursion = 10
+
 // CustomList for fastdialer
 type CustomList []string
 
@@ -12,7 +14,7 @@ func (c *CustomList) String() string {
 
 // Set a new global header
 func (c *CustomList) Set(value string) error {
-	values := fileutil.LoadCidrsFromSliceOrFile(value, ",")
+	values := fileutil.LoadCidrsFromSliceOrFileWithMaxRecursion(value, ",", maxRecursion)
 	*c = append(*c, values...)
 	return nil
 }
