@@ -1,11 +1,12 @@
 package customport
 
 import (
+	"fmt"
 	"strconv"
 	"strings"
 
-	"github.com/projectdiscovery/gologger"
-	"github.com/projectdiscovery/sliceutil"
+	"github.com/pkg/errors"
+	sliceutil "github.com/projectdiscovery/utils/slice"
 
 	"github.com/projectdiscovery/httpx/common/httpx"
 )
@@ -52,39 +53,38 @@ func (c *CustomPorts) Set(value string) error {
 		potentialRange := strings.Split(potentialPort, "-")
 		// it's a single port?
 		if len(potentialRange) < portRangeParts {
-			if p, err := strconv.Atoi(potentialPort); err == nil {
-				if existingProtocol, ok := Ports[p]; ok {
-					if existingProtocol == httpx.HTTP && protocol == httpx.HTTPS || existingProtocol == httpx.HTTPS && protocol == httpx.HTTP {
-						protocol = httpx.HTTPandHTTPS
-					}
-				}
-				Ports[p] = protocol
-			} else {
-				gologger.Warning().Msgf("Could not cast port to integer, your value: %s, resulting error %s. Skipping it\n",
-					potentialPort, err.Error())
+			p, err := strconv.Atoi(potentialPort)
+			if err != nil {
+				return errors.Wrap(err, fmt.Sprintf("Could not cast port to integer from your value: %s\n", potentialPort))
 			}
+			if err := checkPortValue(p); err != nil {
+				return err
+			}
+			if existingProtocol, ok := Ports[p]; ok {
+				if existingProtocol == httpx.HTTP && protocol == httpx.HTTPS || existingProtocol == httpx.HTTPS && protocol == httpx.HTTP {
+					protocol = httpx.HTTPandHTTPS
+				}
+			}
+			Ports[p] = protocol
 		} else {
 			// expand range
-			var lowP, highP int
 			lowP, err := strconv.Atoi(potentialRange[0])
 			if err != nil {
-				gologger.Warning().Msgf("Could not cast first port of your port range(%s) to integer, your value: %s, resulting error %s. Skipping it\n",
-					potentialPort, potentialRange[0], err.Error())
-				continue
+				return errors.Wrap(err, fmt.Sprintf("Could not cast first port of your range(%s) to integer from your value: %s", potentialPort, potentialRange[0]))
 			}
-			highP, err = strconv.Atoi(potentialRange[1])
+			if err := checkPortValue(lowP); err != nil {
+				return errors.Wrap(err, fmt.Sprintf("first port of your range(%d)", lowP))
+			}
+			highP, err := strconv.Atoi(potentialRange[1])
 			if err != nil {
-				gologger.Warning().Msgf("Could not cast last port of your port range(%s) to integer, "+
-					"your value: %s, resulting error %s. Skipping it\n",
-					potentialPort, potentialRange[1], err.Error())
-				continue
+				return errors.Wrap(err, fmt.Sprintf("Could not cast last port of your port range(%s) to integer from your value: %s", potentialPort, potentialRange[1]))
+			}
+			if err := checkPortValue(highP); err != nil {
+				return errors.Wrap(err, fmt.Sprintf("last port of your range(%d)", lowP))
 			}
 
 			if lowP > highP {
-				gologger.Warning().Msgf("first value of port range should be lower than the last part port "+
-					"in that range, your range: [%d, %d]. Skipping it\n",
-					lowP, highP)
-				continue
+				return fmt.Errorf("First value of port range should be lower than the last port from your range: [%d, %d]", lowP, highP)
 			}
 
 			for i := lowP; i <= highP; i++ {
@@ -99,5 +99,12 @@ func (c *CustomPorts) Set(value string) error {
 	}
 
 	*c = append(*c, value)
+	return nil
+}
+
+func checkPortValue(port int) error {
+	if port > 65535 {
+		return errors.New("port value is bigger than 65535")
+	}
 	return nil
 }
