@@ -1,10 +1,12 @@
 package httpx
 
 import (
-	"net/http"
+	"bytes"
 	"strings"
 
+	"github.com/PuerkitoBio/goquery"
 	"github.com/projectdiscovery/httpx/common/slice"
+	stringsutil "github.com/projectdiscovery/utils/strings"
 )
 
 // CSPHeaders is an incomplete list of most common CSP headers
@@ -21,22 +23,28 @@ type CSPData struct {
 }
 
 // CSPGrab fills the CSPData
-func (h *HTTPX) CSPGrab(r *http.Response) *CSPData {
+func (h *HTTPX) CSPGrab(r *Response) *CSPData {
 	domains := make(map[string]struct{})
+	// extract from headers
 	for _, cspHeader := range CSPHeaders {
-		cspRaw := r.Header.Get(cspHeader)
-		if cspRaw != "" {
-			rules := strings.Split(cspRaw, ";")
-			for _, rule := range rules {
-				// rule is like aa bb domain1 domain2 domain3
-				tokens := strings.Split(rule, " ")
-				// we extracts only potential domains
-				for _, t := range tokens {
-					if isPotentialDomain(t) {
-						domains[t] = struct{}{}
+		if cspValues, ok := r.Headers[cspHeader]; ok {
+			for _, cspValue := range cspValues {
+				parsePotentialDomains(domains, cspValue)
+			}
+		}
+	}
+
+	// extract from body
+	if len(r.Data) > 0 {
+		doc, err := goquery.NewDocumentFromReader(bytes.NewReader(r.Data))
+		if err == nil {
+			doc.Find("meta").Each(func(i int, s *goquery.Selection) {
+				if _, ok := s.Attr("http-equiv"); ok {
+					if content, ok := s.Attr("content"); ok {
+						parsePotentialDomains(domains, content)
 					}
 				}
-			}
+			})
 		}
 	}
 
@@ -44,6 +52,17 @@ func (h *HTTPX) CSPGrab(r *http.Response) *CSPData {
 		return &CSPData{Domains: slice.ToSlice(domains)}
 	}
 	return nil
+}
+
+func parsePotentialDomains(domains map[string]struct{}, data string) {
+	// rule is like aa bb domain1 domain2 domain3
+	tokens := stringsutil.SplitAny(data, " ", ";", ",")
+	// we extracts only potential domains
+	for _, t := range tokens {
+		if isPotentialDomain(t) {
+			domains[t] = struct{}{}
+		}
+	}
 }
 
 func isPotentialDomain(s string) bool {
