@@ -4,8 +4,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/projectdiscovery/cryptoutil"
-	"github.com/projectdiscovery/httputil"
+	"github.com/projectdiscovery/tlsx/pkg/tlsx/clients"
+	httputil "github.com/projectdiscovery/utils/http"
 )
 
 // Response contains the response to a server
@@ -18,7 +18,7 @@ type Response struct {
 	RawHeaders    string
 	Words         int
 	Lines         int
-	TLSData       *cryptoutil.TLSData
+	TLSData       *clients.Response
 	CSPData       *CSPData
 	HTTP2         bool
 	Pipeline      bool
